Rename check_hip.h to CheckHip.h
:Release Notes:

:Detailed Notes:

:Testing Performed:
No tests

:QA Notes:

:Issues Addressed:
